update to use "platform" config for io or html
