+++ lib/socketcluster_client.dart
@@ -1,12 +1,3 @@
 library socketcluster_client;
 
-import 'dart:io';
-import 'dart:async';
-import 'dart:convert';
-
-part 'src/socketcluster_client.dart';
-part 'src/emitter.dart';
-part 'src/parser.dart';
-part 'src/channel.dart';
-part 'src/reconnect_strategy.dart';
-part 'src/basic_listener.dart';
+export 'package:socketcluster_client/src/socketcluster_client.dart';
